Moving SearchError to types sub-module
