--- openghg/exceptions/__init__.py
@@ -1,8 +1,0 @@
-
-
-class OpenGHGError(Exception):
-    pass
-
-
-class SearchError(OpenGHGError):
-    pass
