+++ openghg/types/_errors.py
@@ -1,20 +1,25 @@
-class InvalidSiteError(Exception):
+
+class OpenGHGError(Exception):
+    """Top level OpenGHG error"""
+
+
+class InvalidSiteError(OpenGHGError):
     """Raised if an invalid site is passed"""
 
 
-class UnknownDataError(Exception):
+class UnknownDataError(OpenGHGError):
     """Raised if an unknown data type is passed"""
 
 
-class FunctionError(Exception):
+class FunctionError(OpenGHGError):
     """Raised if a serverless function cannot be called correctly"""
 
 
-class ObjectStoreError(Exception):
+class ObjectStoreError(OpenGHGError):
     """Raised if an error accessing an object at a key in the object store occurs"""
 
 
-class DatasourceLookupError(Exception):
+class DatasourceLookupError(OpenGHGError):
     """Raised if Datasource lookup fails"""
 
 
@@ -22,9 +27,13 @@ class EncodingError(ObjectStoreError):
     pass
 
 
-class MutexTimeoutError(Exception):
+class MutexTimeoutError(OpenGHGError):
     pass
 
 
-class RequestBucketError(Exception):
+class RequestBucketError(OpenGHGError):
     pass
+
+
+class SearchError(OpenGHGError):
+    """Related to searching the object store"""
